Fix action use on enterprise/self-hosted instances
Noticed a hard coded `github.com` that broke the use of this action on a Github Enterprise instance, with the error:
```
(node:709008) UnhandledPromiseRejectionWarning: HttpError: Bad credentials
    at /home/azureadmin/actions-runner/_work/_actions/patrickpaulin/ms-teams-deploy-card/master/dist/main/index.js:1:462986
    at processTicksAndRejections (internal/process/task_queues.js:93:5)
    at async getOctokitCommit (/home/azureadmin/actions-runner/_work/_actions/patrickpaulin/ms-teams-deploy-card/master/dist/main/index.js:1:37012)
    at async Object.formatAndNotify (/home/azureadmin/actions-runner/_work/_actions/patrickpaulin/ms-teams-deploy-card/master/dist/main/index.js:1:37508)
(node:709008) UnhandledPromiseRejectionWarning: Unhandled promise rejection. This error originated either by throwing inside of an async function without a catch block, or by rejecting a promise which was not handled with .catch(). (rejection id: 1)
(node:709008) [DEP0018] DeprecationWarning: Unhandled promise rejections are deprecated. In the future, promise rejections that are not handled will terminate the Node.js process with a non-zero exit code.```

The credentials are not faulty, but assuming that the repository is always available at github.com is.